requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
djangorestframework==3.11.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
